requirements.txt
fastapi 
uvicorn[standard]
pandas
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
python-dotenv
ipykernel
psycopg2-binary